Rename hodge.js to circularity.js
The file no longer uses Helmholtz-Hodge decomposition — it implements
cycle flow decomposition. Update imports in data.js and test file.
